# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/sales_teams:
    get:
      tags:
        - SalesTeams
      summary: List SalesTeams
      description: Get a paginated list of SalesTeams with optional filtering and sorting
      operationId: listSalesTeams
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesTeamList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalesTeams
      summary: Create SalesTeam
      description: Create a new SalesTeam
      operationId: createSalesTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalesTeamRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesTeam'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/sales_person_allocations:
    get:
      tags:
        - SalesPersonAllocations
      summary: List SalesPersonAllocations
      description: Get a paginated list of SalesPersonAllocations with optional filtering and sorting
      operationId: listSalesPersonAllocations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesPersonAllocationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalesPersonAllocations
      summary: Create SalesPersonAllocation
      description: Create a new SalesPersonAllocation
      operationId: createSalesPersonAllocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalesPersonAllocationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesPersonAllocation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    SalesTeam:
      type: object
      required:
        - id
        - company_id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        description:
          nullable: true
          type: string
          maxLength: 300
        metadata:
          type: object
...
    CreateSalesTeamRequest:
      type: object
      required:
        - company_id
        - name
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        description:
          nullable: true
          type: string
          maxLength: 300
        metadata:
          type: object
...
    SalesTeamList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesTeam'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    SalesPersonAllocation:
      type: object
      required:
        - id
        - order_id
        - company_id
        - sales_person_id
        - allocated_pct
        - metadata
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        sales_person_id:
          type: string
          format: uuid
        sales_team_id:
          nullable: true
          type: string
          format: uuid
        allocated_pct:
          type: string
        metadata:
          type: object
...
    CreateSalesPersonAllocationRequest:
      type: object
      required:
        - order_id
        - company_id
        - sales_person_id
        - allocated_pct
        - metadata
      properties:
        order_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        sales_person_id:
          type: string
          format: uuid
        sales_team_id:
          nullable: true
          type: string
          format: uuid
        allocated_pct:
          type: string
        metadata:
          type: object
...
    SalesPersonAllocationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesPersonAllocation'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []