# src/presentation/http/sales_person_allocation_handler.rs
//! SalesPersonAllocation REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{SalesPersonAllocationService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateSalesPersonAllocationDto, UpdateSalesPersonAllocationDto, PatchSalesPersonAllocationDto, SalesPersonAllocationResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum SalesPersonAllocationError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for SalesPersonAllocationError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for SalesPersonAllocationError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "SALESPERSONALLOCATION_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "SALESPERSONALLOCATION_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "SALESPERSONALLOCATION_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "SALESPERSONALLOCATION_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for SalesPersonAllocation.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /sales_person_allocations | List with pagination |
/// | POST | /sales_person_allocations | Create new |
/// | GET | /sales_person_allocations/:id | Get by ID |
/// | PUT | /sales_person_allocations/:id | Full update |
/// | PATCH | /sales_person_allocations/:id | Partial update |
/// | DELETE | /sales_person_allocations/:id | Soft delete |
/// | POST | /sales_person_allocations/bulk | Bulk create |
/// | POST | /sales_person_allocations/upsert | Upsert |
/// | GET | /sales_person_allocations/trash | List deleted |
/// | POST | /sales_person_allocations/:id/restore | Restore |
/// | DELETE | /sales_person_allocations/empty | Empty trash |
/// | GET | /sales_person_allocations/:id/deleted | Get deleted by ID |
/// | DELETE | /sales_person_allocations/trash/:id | Permanent delete from trash |
/// | GET | /sales_person_allocations/count | Count active entities |
/// | GET | /sales_person_allocations/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(SalesPersonAllocationService::with_repository(repository));
/// let router = create_sales_person_allocation_routes(service);
/// ```
pub fn create_sales_person_allocation_routes(service: Arc<SalesPersonAllocationService>) -> Router {
    BackboneCrudHandler::<SalesPersonAllocationService, SalesPersonAllocation, CreateSalesPersonAllocationDto, UpdateSalesPersonAllocationDto, SalesPersonAllocationResponseDto>::routes(
        service,
        "/sales_person_allocations",
    )
}

/// Create Axum router with only the read (GET) endpoints for SalesPersonAllocation.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_sales_person_allocation_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_sales_person_allocation_read_routes(service: Arc<SalesPersonAllocationService>) -> Router {
    BackboneCrudHandler::<SalesPersonAllocationService, SalesPersonAllocation, CreateSalesPersonAllocationDto, UpdateSalesPersonAllocationDto, SalesPersonAllocationResponseDto>::read_routes(
        service,
        "/sales_person_allocations",
    )
}

/// Create Axum router with only the write (mutation) endpoints for SalesPersonAllocation.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_sales_person_allocation_write_routes(service: Arc<SalesPersonAllocationService>) -> Router {
    BackboneCrudHandler::<SalesPersonAllocationService, SalesPersonAllocation, CreateSalesPersonAllocationDto, UpdateSalesPersonAllocationDto, SalesPersonAllocationResponseDto>::write_routes(
        service,
        "/sales_person_allocations",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_sales_person_allocation_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<SalesPersonAllocationService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_sales_person_allocation_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}