# src/presentation/http/sales_team_handler.rs
//! SalesTeam REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{SalesTeamService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateSalesTeamDto, UpdateSalesTeamDto, PatchSalesTeamDto, SalesTeamResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum SalesTeamError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for SalesTeamError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for SalesTeamError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "SALESTEAM_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "SALESTEAM_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "SALESTEAM_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "SALESTEAM_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for SalesTeam.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /sales_teams | List with pagination |
/// | POST | /sales_teams | Create new |
/// | GET | /sales_teams/:id | Get by ID |
/// | PUT | /sales_teams/:id | Full update |
/// | PATCH | /sales_teams/:id | Partial update |
/// | DELETE | /sales_teams/:id | Soft delete |
/// | POST | /sales_teams/bulk | Bulk create |
/// | POST | /sales_teams/upsert | Upsert |
/// | GET | /sales_teams/trash | List deleted |
/// | POST | /sales_teams/:id/restore | Restore |
/// | DELETE | /sales_teams/empty | Empty trash |
/// | GET | /sales_teams/:id/deleted | Get deleted by ID |
/// | DELETE | /sales_teams/trash/:id | Permanent delete from trash |
/// | GET | /sales_teams/count | Count active entities |
/// | GET | /sales_teams/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(SalesTeamService::with_repository(repository));
/// let router = create_sales_team_routes(service);
/// ```
pub fn create_sales_team_routes(service: Arc<SalesTeamService>) -> Router {
    BackboneCrudHandler::<SalesTeamService, SalesTeam, CreateSalesTeamDto, UpdateSalesTeamDto, SalesTeamResponseDto>::routes(
        service,
        "/sales_teams",
    )
}

/// Create Axum router with only the read (GET) endpoints for SalesTeam.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_sales_team_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_sales_team_read_routes(service: Arc<SalesTeamService>) -> Router {
    BackboneCrudHandler::<SalesTeamService, SalesTeam, CreateSalesTeamDto, UpdateSalesTeamDto, SalesTeamResponseDto>::read_routes(
        service,
        "/sales_teams",
    )
}

/// Create Axum router with only the write (mutation) endpoints for SalesTeam.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_sales_team_write_routes(service: Arc<SalesTeamService>) -> Router {
    BackboneCrudHandler::<SalesTeamService, SalesTeam, CreateSalesTeamDto, UpdateSalesTeamDto, SalesTeamResponseDto>::write_routes(
        service,
        "/sales_teams",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_sales_team_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<SalesTeamService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_sales_team_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}